+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Version 1.2.2 (not released yet)
 - change screen-shake emulated FPS to 150 to match most common PF configuration
 - fix bullet collision testing if mover is in the line of the shot
 - optimize Flamethower rendering (client-side)
+- add 'debug' command for enabling RF built-in debugging features
 - stability improvements
 
 Version 1.2.1 (released 2017-08-15)